Added icons and structure
I found some icons to use with the application and added some basic CSS
and HTML to get a basic idea of where things will eventually go. This
includes some divs to separate different parts of the application, a
footer for attribution, and some buttons that will eventually become the
interface for the metronome itself.