Centered the Vertical Line Correctly
Made a small change to the vertical line that displays in the screen.
It was intended to be centered, but then while looking at it I noticed
that it was slightly off. This was due to the fact that I had set up
the line using left positioning. This mean that the edge of the line
would be centered, but not the center of the line. By adding a transform,
the element is shifted back by it's own width (50% in this case) so that
it is now centered properly.