Removed images Changed Styling and Structure
I tried to use a CSS sprite for the background of the different beat
buttons but could not find a single icon set that had all of the icons I
would need. I also realized that the sizing was going to be a problem
unless I used images that I made or found that would not really match
each other. Maybe in the future I could generate my own sprite image
using images taken from a standard notation app. For now, these images
have been replaced by text that I can manipulate later. I also tried to
add some media queries for sizing, but struggled to get anything to look
right on smaller screens. I'll test how it looks now on an Iphone SE and
make adjustments as necessary. I also updated the class names to be in
keeping with the CSS naming convention as opposed to camel casing.