Added Slider functionality, Adjusted ID names
The slider now correctly reads out the output BPM. This is handled with a simple
script that uses the DOM to read the value of the slider and output it in the
BPM span. By utilizing a function (NOT an arrow function because of the use of
"this"), the textContent of the BPM span can be updated every time there is an
input on the slider.